- reduce chances of float <-> double conversions happening

+++ src/pump.h
@@ -129,7 +129,7 @@ struct effect_creator {
 extern int sin_lookup_table[SIN_LOOKUP_SIZE+1];
 
 /* [0 .. 1[ -> sin(0 .. 2pi) */
-#define sin_lookup(scale) ((double) sin_lookup_table[(int) ((scale) * SIN_LOOKUP_SIZE)] / SIN_LOOKUP_AMPLITUDE)
+#define sin_lookup(scale) ((float) sin_lookup_table[(int) ((scale) * SIN_LOOKUP_SIZE)] / SIN_LOOKUP_AMPLITUDE)
 
 extern volatile unsigned short write_track;
